Refactor: 파일 구조 수정 #SCRUM-46
