+++ src/features/minigame/flipcardgame/model/filpCardStore.ts
@@ -1,11 +1,5 @@
 import { create } from 'zustand';
-
-export interface Card {
-  card_id: string;
-  card_title: string;
-  card_content: string;
-  category: string;
-}
+import type { Card } from '../types/cardTypes';
 
 export interface FlipCardState {
   allCards: Card[]; // 전체 카드 데이터
