+++ src/features/minigame/flipcardgame/types/cardTypes.ts
@@ -1,0 +1,6 @@
+export interface Card {
+    card_id: string;
+    card_title: string;
+    card_content: string;
+    category: string;
+  }
